Bug 820559: Workaround old Android gcc build-bustage by casting anonymous enum template args to ints, r=ekr
UltraBlame original commit: 7ef53b8f1ff60cd7f796304cd3dc822ae7a653a2

+++ media/webrtc/signaling/src/peerconnection/PeerConnectionImpl.cpp
@@ -1028,7 +1028,8 @@ PeerConnectionImpl::ChangeReadyState(PeerConnectionImpl::ReadyState aReadyState)
   
   RUN_ON_THREAD(mThread, WrapRunnable(mPCObserver,
                                       &IPeerConnectionObserver::OnStateChange,
-                                      IPeerConnectionObserver::kReadyState),
+                                      
+                                      static_cast<int>(IPeerConnectionObserver::kReadyState)),
     NS_DISPATCH_NORMAL);
 }
 
@@ -1076,7 +1077,8 @@ PeerConnectionImpl::IceGatheringCompleted_m(NrIceCtx *aCtx)
     RUN_ON_THREAD(mThread,
                   WrapRunnable(mPCObserver,
                                &IPeerConnectionObserver::OnStateChange,
-                               IPeerConnectionObserver::kIceState),
+                               
+                               static_cast<int>(IPeerConnectionObserver::kIceState)),
                   NS_DISPATCH_NORMAL);
   }
 #endif
@@ -1107,7 +1109,8 @@ PeerConnectionImpl::IceCompleted_m(NrIceCtx *aCtx)
     RUN_ON_THREAD(mThread,
                   WrapRunnable(mPCObserver,
                                &IPeerConnectionObserver::OnStateChange,
-                               IPeerConnectionObserver::kIceState),
+                               
+			       static_cast<int>(IPeerConnectionObserver::kIceState)),
                   NS_DISPATCH_NORMAL);
   }
 #endif
